Controllerのfrontendにあるやつらlist配下にするのやめた
