+++ app/controllers/frontend/ActivityDetail.php
@@ -1,16 +1,16 @@
 <?php
-namespace app\controllers\frontend\lists;
+namespace app\controllers\frontend;
 
 use Silex\Application;
 use Symfony\Component\HttpFoundation\Request;
 
-class Introduction
+class ActivityDetail
 {
     public function indexAction(Application $app, Request $request){
 
         return $app['twig']->render('index.twig', array(
-            'titleAnimation' => "fadeInDown",
-            'urlCode' => INTRODUCTION_URL,
+            'titleAnimation' => "bounce",
+            'urlCode' => ACTIVITY_DETAIL_URL,
         ));
     }
 }
