+++ app/controllers/frontend/ClanMember.php
@@ -1,16 +1,16 @@
 <?php
-namespace app\controllers\frontend\lists;
+namespace app\controllers\frontend;
 
 use Silex\Application;
 use Symfony\Component\HttpFoundation\Request;
 
-class JoinToClan
+class ClanMember
 {
     public function indexAction(Application $app, Request $request){
 
         return $app['twig']->render('index.twig', array(
-            'titleAnimation' => "swing",
-            'urlCode' => JOIN_TO_CLAN_URL,
+            'titleAnimation' => "rubberBand",
+            'urlCode' => CLAN_MEMBER_URL,
         ));
     }
 }
