+++ app/controllers/frontend/Gallery.php
@@ -1,16 +1,16 @@
 <?php
-namespace app\controllers\frontend\lists;
+namespace app\controllers\frontend;
 
 use Silex\Application;
 use Symfony\Component\HttpFoundation\Request;
 
-class ClanMember
+class Gallery
 {
     public function indexAction(Application $app, Request $request){
 
         return $app['twig']->render('index.twig', array(
-            'titleAnimation' => "rubberBand",
-            'urlCode' => CLAN_MEMBER_URL,
+            'titleAnimation' => "tada",
+            'urlCode' => GALLERY_URL,
         ));
     }
 }
