+++ app/controllers/frontend/Introduction.php
@@ -1,16 +1,16 @@
 <?php
-namespace app\controllers\frontend\lists;
+namespace app\controllers\frontend;
 
 use Silex\Application;
 use Symfony\Component\HttpFoundation\Request;
 
-class Gallery
+class Introduction
 {
     public function indexAction(Application $app, Request $request){
 
         return $app['twig']->render('index.twig', array(
-            'titleAnimation' => "tada",
-            'urlCode' => GALLERY_URL,
+            'titleAnimation' => "fadeInDown",
+            'urlCode' => INTRODUCTION_URL,
         ));
     }
 }
